+++ CHANGELOG.md
@@ -9,6 +9,10 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Added
 
+- Support for [`NLog`](https://www.nuget.org/packages/NLog/)
+  logs instrumentation for versions `5.*` and `6.*` on .NET using duck typing
+  for zero-config auto-injection.
+
 ### Changed
 
 #### Dependency updates